// src/enums/ActivityAction.ts
export enum ActivityAction {
  // Cases
  CASE_CREATED = 'CASE_CREATED',
  CASE_UPDATED = 'CASE_UPDATED',
  CASE_ASSIGNED = 'CASE_ASSIGNED',
  CASE_COMPLETED = 'CASE_COMPLETED',
  CASE_CANCELLED = 'CASE_CANCELLED',
  STATUS_CHANGED = 'STATUS_CHANGED',

  // Files
  FILE_UPLOADED = 'FILE_UPLOADED',
  FILE_UPDATED = 'FILE_UPDATED',
  FILE_DELETED = 'FILE_DELETED',

  // Assignment
  ASSIGNMENT_CREATED = 'ASSIGNMENT_CREATED',
  ASSIGNMENT_ACCEPTED = 'ASSIGNMENT_ACCEPTED',
  ASSIGNMENT_REJECTED = 'ASSIGNMENT_REJECTED',
  ASSIGNMENT_SUBMITTED = 'ASSIGNMENT_SUBMITTED',

  // Payment
  PAYMENT_CREATED = 'PAYMENT_CREATED',
  PAYMENT_APPROVED = 'PAYMENT_APPROVED',
  PAYMENT_REJECTED = 'PAYMENT_REJECTED',

  // Subscription
  SUBSCRIPTION_CREATED = 'SUBSCRIPTION_CREATED',
  SUBSCRIPTION_RENEWED = 'SUBSCRIPTION_RENEWED',
  SUBSCRIPTION_CANCELLED = 'SUBSCRIPTION_CANCELLED',

  // Plans
  PLAN_CREATED = 'PLAN_CREATED',
  PLAN_UPDATED = 'PLAN_UPDATED',
  PLAN_DELETED = 'PLAN_DELETED',
}